1.修改了获取定位时使用data class 的错误，原因是在关闭定位时会导致一些原来以为不为空的值为空，但是kotlin并没有报错，现在全部允许为空
2.优化了新增位置界面

+++ flutter/crowdsourcing/lib/pages/LocationPage.dart
@@ -53,11 +53,18 @@ class LocationPage extends StatelessWidget {
                                   locations[index].phone,
                                 ),
                                 GestureDetector(
-                                  child: Icon(Icons.delete_outline),
-                                  onTap: () {
-                                    locationModel
-                                        .deleteLocation(locations[index].id);
-                                    _setState(() {});
+                                  behavior: HitTestBehavior.opaque,
+                                  child: Container(
+                                      width: 40,
+                                      child: Icon(Icons.delete_outline)),
+                                  onTap: () async {
+                                    bool delete =
+                                        await showDeleteConfirmDialog1(context);
+                                    if (delete) {
+                                      locationModel
+                                          .deleteLocation(locations[index].id);
+                                      _setState(() {});
+                                    }
                                   },
                                 )
                               ],
@@ -91,5 +98,31 @@ class LocationPage extends StatelessWidget {
                 });
           });
         })));
+  }
+
+  // 弹出对话框
+  Future<bool> showDeleteConfirmDialog1(context) {
+    return showDialog<bool>(
+      context: context,
+      builder: (context) {
+        return AlertDialog(
+          title: Text("提示"),
+          content: Text("您确定要删除当前位置吗?"),
+          actions: <Widget>[
+            FlatButton(
+              child: Text("取消"),
+              onPressed: () => Navigator.of(context).pop(false), // 关闭对话框
+            ),
+            FlatButton(
+              child: Text("删除"),
+              onPressed: () {
+                //关闭对话框并返回true
+                Navigator.of(context).pop(true);
+              },
+            ),
+          ],
+        );
+      },
+    );
   }
 }
